Spectra MS backend storing data in a MongoDb database

MsBackendMongoDb is an implementation of the Spectra::MsBackend() class for Spectra::Spectra() objects, storing and retrieving MS data from a MongoDb database.

This backend allows read-only access to spectra and peaks stored in MongoDb, but supports temporary modifications of spectra variables via caching.

# S4 method for class 'MsBackendMongoDb'
show(object)

# S4 method for class 'MsBackendMongoDb'
backendInitialize(object, dbcon, data, ...)

# S4 method for class 'MsBackendMongoDb'
dataStorage(object)

# S4 method for class 'MsBackendMongoDb'
peaksData(object, columns = c("mz", "intensity"))

# S4 method for class 'MsBackendMongoDb'
peaksVariables(object)

# S4 method for class 'MsBackendMongoDb'
mz(object)

# S4 method for class 'MsBackendMongoDb'
mz(object) <- value

# S4 method for class 'MsBackendMongoDb'
intensity(object)

# S4 method for class 'MsBackendMongoDb'
intensity(object) <- value

# S4 method for class 'MsBackendMongoDb'
spectraNames(object)

# S4 method for class 'MsBackendMongoDb'
spectraNames(object) <- value

# S4 method for class 'MsBackendMongoDb'
spectraData(object, columns = spectraVariables(object))

# S4 method for class 'MsBackendMongoDb'
x[i, j, ..., drop = FALSE]

# S4 method for class 'MsBackendMongoDb,ANY'
extractByIndex(object, i)

# S4 method for class 'MsBackendMongoDb'
reset(object)

Arguments

object

A MsBackendMongoDb instance

dbcon

list() of MongoDb collections, e.g., list(compounds_info_coll = mongo_conn)

data

Optional DataFrame with data to insert into MongoDb

...

Additional arguments passed internally

columns

Character() vector of columns to fetch for peaksData or spectraData

value

Replacement values for setter functions

x

for [i]: a MsBackendMongoDb instance

i

integer() Indices to subset the backend

j

integer() Ignored.

drop

logical(1), ignored for subsetting

Value

Depends on the method. See individual function documentation.

Note

MsBackendMongoDb keeps a connection to a MongoDb collection(s) and fetches spectra and peaks data on demand. By storing only the primary keys (spectrum_id) in memory, it ensures minimal memory usage, while peaks and metadata are fetched dynamically.

The backend inherits from Spectra::MsBackendCached() and supports temporary modification of spectra variables via the $<- operator. Original data in the database cannot be changed (intensities and m/z values are read-only).

Creation of backend objects

New backend objects can be created using:

backend <- MsBackendMongoDb(dbcon = list(compounds_info_coll = mongo_connection))
backend <- backendInitialize(backend, dbcon = backend@dbcon)

dbcon: A list of MongoDb collections (e.g. from mongolite::mongo()) as created by the connectMsBackendMongoDb().

backendInitialize(): populates the object with spectrum IDs and caches metadata. If data is provided, it can insert data into the collection.

Accessing data

peaksData(object, columns = c("mz", "intensity")): returns a list of matrices containing the peak data for each spectrum.

peaksVariables(object): returns available columns for peaksData.

intensity(object), mz(object): return SimpleList objects with numeric intensity or m/z values for each spectrum.

spectraData(object, columns): returns a DataFrame with requested spectrum metadata columns.

spectraNames(object): returns spectrum_id values as character.

Read-only data

intensity<- and mz<- are not supported.

spectraNames<- is not supported.

Subsetting and extraction

[i] and extractByIndex(object, i) subset the backend by spectrum indices. Original data in the database is never changed.

reset(object) restores the backend to its original state, removing subsetting and cached variables.

Miscellaneous

dataStorage(object) returns a string describing the MongoDb collections used by the backend.

Implementation notes

Internally, the backend stores:

spectraIds: integer vector of spectrum IDs

dbcon: MongoDb connection(s)

.collections: list of collection names

peak_fun: function used to fetch peak data

localData: cached spectrum variables

nspectra: total number of spectra

Data access functions use peak_fun to fetch peaks on demand.

Author

Ahlam Mentag